refactor: Centralize changes in enrollment_switchio app
